Add gid wrapper for mock
This commit enables running mock for users in the retrace group without being
in the mock group.

The original retrace-server-interact is moved into /usr/libexec and in its
place a gid wrapper is placed. This wrapper check if the executing user is in
the retrace group and then change group identification to mock. Then the
original r-s-i can be executed.

This commit does not change user interface, that means using r-s-i from users
point of view does not change.

Signed-off-by: Matej Marusak <mmarusak@redhat.com>